Yade

+++ sos-webservices/sos-webservices-json/src/main/java/com/sos/joc/model/yade/TransferFilesSummary.java
@@ -16,69 +16,69 @@ import org.apache.commons.lang.builder.ToStringBuilder;
 public class TransferFilesSummary {
 
     /**
-     * non negative integer
+     * non negative long
      * <p>
      * 
      * (Required)
      * 
      */
     @JsonProperty("successful")
-    private Integer successful;
+    private Long successful;
     /**
-     * non negative integer
+     * non negative long
      * <p>
      * 
      * (Required)
      * 
      */
     @JsonProperty("failed")
-    private Integer failed;
+    private Long failed;
 
     /**
-     * non negative integer
+     * non negative long
      * <p>
      * 
      * (Required)
      * 
      */
     @JsonProperty("successful")
-    public Integer getSuccessful() {
+    public Long getSuccessful() {
         return successful;
     }
 
     /**
-     * non negative integer
+     * non negative long
      * <p>
      * 
      * (Required)
      * 
      */
     @JsonProperty("successful")
-    public void setSuccessful(Integer successful) {
+    public void setSuccessful(Long successful) {
         this.successful = successful;
     }
 
     /**
-     * non negative integer
+     * non negative long
      * <p>
      * 
      * (Required)
      * 
      */
     @JsonProperty("failed")
-    public Integer getFailed() {
+    public Long getFailed() {
         return failed;
     }
 
     /**
-     * non negative integer
+     * non negative long
      * <p>
      * 
      * (Required)
      * 
      */
     @JsonProperty("failed")
-    public void setFailed(Integer failed) {
+    public void setFailed(Long failed) {
         this.failed = failed;
     }
